fixed check in cleaning, hopefully

+++ location/models.py
@@ -40,22 +40,25 @@ class Person(models.Model):
                 checkIn.delete()
             elif checkIn == newCheckIn:
                 continue
+            elif checkIn.touches(newCheckIn, assertActive = True):
+                if checkIn.place == newCheckIn.place:
+                    print(f"merging touching checkins: {checkIn} and {newCheckIn}")
+                    if verbose:
+                        self.send(f"Merging {self.name}'s check ins at {checkIn.prettyNoName()} "
+                                    f"and {newCheckIn.prettyNoName()}.")
+                    newCheckIn.start_time = min(newCheckIn.start_time, checkIn.start_time)
+                    newCheckIn.end_time = max(newCheckIn.end_time, checkIn.end_time)
+                    newCheckIn.clean()
+                    newCheckIn.save()
+                    checkIn.delete()
+                else:
+                    print("two checkins touching, letting it pass")
             elif checkIn.overlaps(newCheckIn):
                 print(f"deleting overlapping checkin: {checkIn}")
                 if verbose:
                     self.send(f"Checking {self.name} out from {checkIn.prettyNoName()} "
                             f"because this overlaps with {newCheckIn.prettyNoName()}.")
                 checkIn.scratch()
-            elif checkIn.touches(newCheckIn, assertActive = True):
-                print(f"merging touching checkins: {checkIn} and {newCheckIn}")
-                if verbose:
-                    self.send(f"Merging {self.name}'s check ins at {checkIn.prettyNoName()} "
-                                f"and {newCheckIn.prettyNoName()}.")
-                newCheckIn.start_time = min(newCheckIn.start_time, checkIn.start_time)
-                newCheckIn.end_time = max(newCheckIn.end_time, checkIn.end_time)
-                newCheckIn.clean()
-                newCheckIn.save()
-                checkIn.delete()
 
     def send(self, outMsg, msgType = "RESPONSE"):
         print("OUT:", outMsg)
@@ -204,7 +207,7 @@ class CheckIn(models.Model):
         return self.start_time < other.end_time and self.end_time > other.start_time
 
     def touches(self, other, assertActive = False) -> bool:
-        if self.scratched or other.scratched:
+        if (self.scratched or other.scratched) and assertActive:
             return False
 
         minTimeGap = min(
